Refactored code for the hyperfluxfno, now showing good performance on the cubic conservation law dataset

+++ src/context_flux_no/models/multiphysics/disco/disco.py
@@ -104,7 +104,7 @@ class DISCO(AbstractMultiphysicsOperator):
         embedding_dim: int,
         patch_size: int,
         num_hypernet_blocks: int,
-        droppath: int,
+        droppath: float,
         num_hypernet_heads: int,
         mlp_hidden_dim: int,
         boundary_condition: str,
